+++ chat-widget/src/components/webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/typingIndicatorMiddleware.tsx
@@ -7,6 +7,7 @@
 import React, { Dispatch, useCallback } from "react";
 
 import { DirectLineSenderRole } from "../../enums/DirectLineSenderRole";
+import { FacadeChatSDK } from "../../../../../common/facades/FacadeChatSDK";
 import { ILiveChatWidgetAction } from "../../../../../contexts/common/ILiveChatWidgetAction";
 import { ILiveChatWidgetContext } from "../../../../../contexts/common/ILiveChatWidgetContext";
 import { ILiveChatWidgetLocalizedTexts } from "../../../../../contexts/common/ILiveChatWidgetLocalizedTexts";
@@ -16,14 +17,15 @@ import { defaultTypingIndicatorBubbleStyles } from "./defaultStyles/defaultTypin
 import { defaultTypingIndicatorContainerStyles } from "./defaultStyles/defaultTypingIndicatorContainerStyles";
 import { defaultTypingIndicatorMessageStyles } from "./defaultStyles/defaultTypingIndicatorMessageStyles";
 import { useChatContextStore } from "../../../../..";
-import useChatSDKStore from "../../../../../hooks/useChatSDKStore";
+import useFacadeSDKStore from "../../../../../hooks/useFacadeChatSDKStore";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const TypingIndicator = ({ activeTyping, visible }: any) => {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const chatSDK: any = useChatSDKStore();
+    
+    const [facadeChatSDK]: [FacadeChatSDK, (facadeChatSDK: FacadeChatSDK) => void] = useFacadeSDKStore();
     const [state, ]: [ILiveChatWidgetContext, Dispatch<ILiveChatWidgetAction>] = useChatContextStore();
-    const debounceTyping = useCallback(debounceLeading(() => chatSDK?.sendTypingEvent()), []);
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const debounceTyping = useCallback(debounceLeading(() => (facadeChatSDK as any)?.sendTypingEvent()), []);
 
     if (!activeTyping || Object.keys(activeTyping).length === 0 || (state.domainStates.liveChatConfig?.LiveChatVersion === 1 && !visible)) {
         return null;